created repositories

+++ Catalog.Domain/Common/Helpers/PagedList.cs
@@ -7,10 +7,10 @@ public class PagedList<T> : List<T>
     public int CurrentPage { get; private set; }
     public int TotalPages { get; private set; }
     public int PageSize { get; private set; }
-    public int TotalCount { get; private set; }
+    public long TotalCount { get; private set; }
     public string? NextCursor { get; private set; }
 
-    public PagedList(List<T> items, int count, int pageNumber, int pageSize, string? nextCursor = null)
+    public PagedList(List<T> items, long count, int pageNumber, int pageSize, string? nextCursor = null)
     {
         TotalCount = count;
         PageSize = pageSize;
